se aplico la arquitectura MVVM
