+++ app/src/main/java/com/example/climasync/view/MainActivity.kt
@@ -1,11 +1,8 @@
-package com.example.climasync
+package com.example.climasync.view
 
 import android.os.Bundle
 import android.view.View
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.example.climasync.databinding.ActivityMainBinding
 import com.example.climasync.utils.FragmentCommunicator
