+++ app/src/main/java/com/example/climasync/view/Personal_information.kt
@@ -1,4 +1,4 @@
-package com.example.climasync
+package com.example.climasync.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,25 +6,27 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
-import com.example.climasync.databinding.FragmentRestorePasswordBinding
+import com.example.climasync.R
+import com.example.climasync.databinding.FragmentPersonalInformationBinding
 import com.example.climasync.utils.FragmentCommunicator
 
 
 /**
 
 A simple [Fragment] subclass.
-Use the [restore_password.newInstance] factory method to
+Use the [personal_information.newInstance] factory method to
 create an instance of this fragment.*/
-class restore_password : Fragment() {
+class personal_information : Fragment() {
     // TODO: Rename and change types of parameters
 
-    private var _binding:FragmentRestorePasswordBinding? = null
-    private val binding get()=_binding!!
+    private var _binding: FragmentPersonalInformationBinding? = null
+    private val binding get() = _binding!!
     private lateinit var communicator: FragmentCommunicator
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
+
         }
     }
 
@@ -32,18 +34,15 @@ class restore_password : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentRestorePasswordBinding.inflate(inflater, container, false)
+        _binding = FragmentPersonalInformationBinding.inflate(inflater, container, false)
         communicator = requireActivity() as MainActivity
         setupView()
-        return (binding.root)
+        return binding.root
     }
 
-    private fun setupView(){
-        binding.flechaRestorePassword.setOnClickListener {
-            findNavController().navigate(R.id.action_restore_password_to_FirstFragment)
-        }
-        binding.resertButton.setOnClickListener {
-            findNavController().navigate(R.id.action_restore_password_to_FirstFragment)
+    private fun setupView() {
+        binding.personalButton.setOnClickListener{
+            findNavController().navigate(R.id.action_personal_information_to_layoutPermission)
         }
     }
