+++ app/src/main/java/com/example/climasync/view/SecondFragment.kt
@@ -1,4 +1,4 @@
-package com.example.climasync
+package com.example.climasync.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import com.example.climasync.R
 import com.example.climasync.databinding.FragmentSecondBinding
 import com.example.climasync.utils.FragmentCommunicator
