+++ app/src/main/res/layout/activity_main2.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".MainActivity2">
+    tools:context=".view.MainActivity2">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_height="wrap_content"
