+++ app/src/main/res/layout/fragment_layout_permission.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="match_parent"
-    tools:context=".LayoutPermission">
+    tools:context=".view.LayoutPermission">
 
     <!-- TODO: Update blank fragment layout -->
