+++ app/src/main/res/layout/fragment_second.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFF9F9"
-    tools:context=".SecondFragment" >
+    tools:context=".view.SecondFragment" >
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
