+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.climasync.FirstFragment"
+        android:name="com.example.climasync.view.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first" >
 
@@ -20,7 +20,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.climasync.SecondFragment"
+        android:name="com.example.climasync.view.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second" >
 
@@ -33,7 +33,7 @@
     </fragment>
     <fragment
         android:id="@+id/restore_password"
-        android:name="com.example.climasync.restore_password"
+        android:name="com.example.climasync.view.restore_password"
         android:label="restore_password" >
         <action
             android:id="@+id/action_restore_password_to_FirstFragment"
@@ -41,7 +41,7 @@
     </fragment>
     <fragment
         android:id="@+id/personal_information"
-        android:name="com.example.climasync.personal_information"
+        android:name="com.example.climasync.view.personal_information"
         android:label="personal_information" >
         <action
             android:id="@+id/action_personal_information_to_layoutPermission"
@@ -49,7 +49,7 @@
     </fragment>
     <fragment
         android:id="@+id/layoutPermission"
-        android:name="com.example.climasync.LayoutPermission"
+        android:name="com.example.climasync.view.LayoutPermission"
         android:label="fragment_layout_permission"
         tools:layout="@layout/fragment_layout_permission" />
 </navigation>
